app_final/requirements.txt

fastapi
uvicorn[standard]
python-dotenv

psycopg2-binary
pgvector
SQLAlchemy

pandas
numpy
sentence-transformers
torch
torchvision

langchain
langchain-core

requests
